refactor: move pages into new dir
