+++ src/old/lib/components/Errors/Page404.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Typography } from '@material-ui/core';
 import { useTranslation } from 'react-i18next';
-import { PageTitle } from '../Pages/PageTitle';
+import { PageTitle } from 'components/Page/PageTitle';
 import { setGALocation } from '../../../../utilities/setGALocation';
 import { langTokens } from '../../../../locales/localizationInit';
