+++ src/views/postCreation/VideoCreation/VideoPostCreation.tsx
@@ -4,6 +4,7 @@ import { useSelector } from 'react-redux';
 import _ from 'lodash';
 import { Box, TextField, Typography } from '@material-ui/core';
 import { useTranslation } from 'react-i18next';
+import { PageTitle } from 'components/Page/PageTitle';
 import {
   setPostDirections,
   setPostOrigin,
@@ -33,7 +34,6 @@ import {
   CreateVideoPostRequestType,
   ExpertResponseType,
 } from '../../../old/lib/utilities/API/types';
-import { PageTitle } from '../../../old/lib/components/Pages/PageTitle';
 import { createPost, getAllExperts } from '../../../old/lib/utilities/API/api';
 import {
   CONTENT_DEBOUNCE_TIMEOUT,
