+++ src/views/postUpdation/TextPostUpdation.tsx
@@ -4,8 +4,8 @@ import _ from 'lodash';
 import { Box, TextField, Typography } from '@material-ui/core';
 import { useTranslation } from 'react-i18next';
 import { DropEvent, FileRejection } from 'react-dropzone';
+import { PageTitle } from 'components/Page/PageTitle';
 import { sanitizeHtml } from '../../old/lib/utilities/sanitizeHtml';
-import { PageTitle } from '../../old/lib/components/Pages/PageTitle';
 import { updatePost, getAllExperts } from '../../old/lib/utilities/API/api';
 import { IDirection, IPost, IOrigin } from '../../old/lib/types';
 import { PostCreationButtons } from '../postCreation/PostCreationButtons';
