+++ src/views/postUpdation/VideoUpdation/VideoPostUpdation.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import _ from 'lodash';
 import { Box, TextField, Typography } from '@material-ui/core';
+import { PageTitle } from 'components/Page/PageTitle';
 import { sanitizeHtml } from '../../../old/lib/utilities/sanitizeHtml';
-import { PageTitle } from '../../../old/lib/components/Pages/PageTitle';
 import { updatePost, getAllExperts } from '../../../old/lib/utilities/API/api';
 import { IDirection, IPost, IOrigin } from '../../../old/lib/types';
 import { PostCreationButtons } from '../../postCreation/PostCreationButtons';
